final version

+++ src/main/java/com/example/application/service/ScheduleService.java
@@ -2,8 +2,12 @@ package com.example.application.service;
 
 import com.example.application.entity.Event;
 import com.example.application.entity.Schedule;
+import com.example.application.entity.User;
+import com.example.application.exception.EventNotFoundException;
+import com.example.application.exception.ScheduleNotFoundException;
 import com.example.application.repository.ScheduleRepository;
 import com.example.application.repository.EventRepository;
+import com.example.application.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
@@ -13,20 +17,19 @@ import java.net.URISyntaxException;
 import java.sql.Date;
 import java.util.ArrayList;
 import java.util.List;
-/*
-* delete
-* update
-* */
+
 @Service
 public class ScheduleService {
 
     private final ScheduleRepository scheduleRepository;
     private final EventRepository eventRepository;
+    private final UserRepository userRepository;
 
     @Autowired
-    public ScheduleService(ScheduleRepository scheduleRepository, EventRepository eventRepository) {
+    public ScheduleService(ScheduleRepository scheduleRepository, EventRepository eventRepository, UserRepository userRepository) {
         this.scheduleRepository = scheduleRepository;
         this.eventRepository = eventRepository;
+        this.userRepository = userRepository;
     }
 
     public List<Long> getSchedulesForMonth(String beginOfMonth, String endOfMonth) {
@@ -47,10 +50,15 @@ public class ScheduleService {
     public List<Event> getTodayEvent(String currentDate) {
         Schedule schedule =
                 scheduleRepository.findByDate(new Date(Long.parseLong(currentDate)));
-        return eventRepository.findByScheduleId(schedule.getId());
+        if(schedule != null) {
+            return eventRepository.findByScheduleId(schedule.getId());
+        } else {
+            throw new ScheduleNotFoundException();
+        }
     }
 
     public ResponseEntity<Event> addEvent(String date, String content) throws URISyntaxException {
+
         Schedule schedule =
                 scheduleRepository.findByDate(new Date(Long.parseLong(date)));
         if(schedule != null) {
@@ -60,8 +68,10 @@ public class ScheduleService {
             eventRepository.save(event);
             return ResponseEntity.created(new URI("/calendar/"+date)).body(event);
         } else {
+            User user = userRepository.findCurrentByEmail();
             schedule = new Schedule();
             schedule.setEvent_date(new Date(Long.parseLong(date)));
+            schedule.setUser(user);
             scheduleRepository.save(schedule);
             Event event = new Event();
             event.setSchedule(schedule);
@@ -71,14 +81,22 @@ public class ScheduleService {
         }
     }
 
-    public void updateEvent(Long id, String content) {
-        Event event = eventRepository.findById(id).get();
-        event.setContent(content);
-        eventRepository.save(event);
+    public ResponseEntity<String> deleteSchedule(String date) {
+        Schedule schedule =
+                scheduleRepository.findByDate(new Date(Long.parseLong(date)));
+        if(schedule != null) {
+            scheduleRepository.delete(schedule);
+        } else {
+            throw new ScheduleNotFoundException();
+        }
+
+        return ResponseEntity.ok("Deleted");
     }
 
-    public void deleteSchedule(String date) {
-        Schedule schedule = scheduleRepository.findByDate(new Date(Long.parseLong(date)));
-        scheduleRepository.delete(schedule);
+    public void updateEvent(Long id, String content) {
+        Event event = eventRepository.findById(id)
+                .orElseThrow(()->new EventNotFoundException());
+        event.setContent(content);
+        eventRepository.save(event);
     }
 }
